backend/requirements.txt
asgiref==3.5.2
cffi==1.15.1
cryptography==38.0.0
Django==4.1.1
django-rest-knox==4.2.0
djangorestframework==3.13.1
pycparser==2.21
python-dotenv==0.21.0
pytz==2022.2.1
sqlparse==0.4.2
